settings-fix
